Make manager API compliant with REST API proposal
