+++ manager/common.py
@@ -1,0 +1,42 @@
+
+import json
+
+import requests
+
+from tornado.web import RequestHandler
+
+VMAAS_URL = "https://webapp-vmaas-stage.1b13.insights.openshiftapps.com"
+DEFAULT_ROUTE = "/api/v1"
+
+
+class BaseHandler(RequestHandler):
+
+    def set_default_headers(self):
+        self.set_header("Access-Control-Allow-Origin", "*")
+        self.set_header("Access-Control-Allow-Headers",
+                        "Content-Type, Access-Control-Allow-Headers, \
+                        Authorization, X-Requested-With")
+        self.set_header('Access-Control-Allow-Methods', 'POST, GET, OPTIONS')
+
+    def options(self):
+        self.set_status(204)
+        self.finish()
+
+
+def vmaas_call(endpoint, data):
+    headers = {'Content-type': 'application/json',
+               'Accept': 'application/json'}
+    response = requests.post(VMAAS_URL + endpoint,
+                             data=json.dumps(data), headers=headers)
+    return response.json()
+
+
+def parse_url(uri, base_handler):
+    prefix = DEFAULT_ROUTE + base_handler
+    if not prefix.endswith("/"):
+        prefix = prefix + "/"
+    if uri.endswith("/"):
+        uri = uri[:-1]
+    request_url = uri[len(prefix):]
+
+    return request_url.split("/")
