+++ manager/main.py
@@ -1,95 +1,27 @@
 #!/usr/bin/python3
 
-import json
 import os
 import sys
 
-import requests
-
-import tornado.httpserver
-from database.database_handler import DatabaseHandler, init_db
-from tornado.ioloop import IOLoop
-from tornado.web import RequestHandler
-
-VMAAS_URL = "https://webapp-vmaas-stage.1b13.insights.openshiftapps.com"
-
-
-class BaseHandler(RequestHandler):
-
-    def set_default_headers(self):
-        self.set_header("Access-Control-Allow-Origin", "*")
-        self.set_header("Access-Control-Allow-Headers",
-                        "Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With")
-        self.set_header('Access-Control-Allow-Methods', 'POST, GET, OPTIONS')
-
-    def post(self):
-        self.write('some post')
-
-    def get(self):
-        self.write('some get')
-
-    def options(self):
-        self.set_status(204)
-        self.finish()
-
-
-class CVEDetailsHandler(BaseHandler):
-    """Handler class returning CVE details."""
-
-    def get(self, *args, **kwargs):
-        synopsis = self.request.uri.split('/')[-1]
-        data = vmaas_call("/api/v1/cves", {"cve_list": [synopsis]})
-        cve_list = data["cve_list"]
-        if synopsis not in cve_list:
-            self.send_error(404)
-        else:
-            self.write(json.dumps(cve_list[synopsis]))
-        self.flush()
-
-    def post(self, *args, **kwargs):
-        self.flush()
-
-
-class AffectedSystemsHandler(BaseHandler):
-    """Handler class returning affected systems by given CVE."""
-
-    def get(self, *args, **kwargs):
-        synopsis = self.request.uri.split('/')[-1]
-        self.conn = DatabaseHandler.get_connection()
-        cur = self.conn.cursor()
-        cur.execute(
-            "SELECT s.uuid, s.display_name, v.cve, v.cvss3_score \
-            FROM system_vulnerabilities v \
-            JOIN system_platform s ON v.uuid = s.uuid \
-            WHERE v.cve = %s", (synopsis,))
-        affected_systems = []
-        for system in cur.fetchall():
-            system_entry = {}
-            system_entry["uuid"] = system[0]
-            system_entry["display_name"] = system[1]
-            system_entry["cve"] = system[2]
-            system_entry["cvss3_score"] = float(system[3])
-            affected_systems.append(system_entry)
-        self.write(json.dumps(affected_systems))
-        self.flush()
-
-    def post(self, *args, **kwargs):
-        self.flush()
+from common import DEFAULT_ROUTE
+from cve_handler import CVEHandler
+from database.database_handler import init_db
+from tornado import httpserver, ioloop, web
 
 
 class ServerApplication():
+
     def __init__(self):
         handlers = [
-            (r"/api/v1/cves/details/CVE-[^/]*", CVEDetailsHandler),
-            (r"/api/v1/cves/affectedsystems/CVE-[^/]*", AffectedSystemsHandler)
+            (DEFAULT_ROUTE + r"/cves/CVE-[^/]*/?[^/]*/?$", CVEHandler)
         ]
-        application = tornado.web.Application(handlers)
-        http_server = tornado.httpserver.HTTPServer(application, ssl_options={
+        application = web.Application(handlers)
+        http_server = httpserver.HTTPServer(application, ssl_options={
             "certfile": os.path.join(sys.path[0], "localhost.crt"),
             "keyfile": os.path.join(sys.path[0], "localhost.key")
         })
         http_server.listen(8443)
-        self.instance = IOLoop.instance()
+        self.instance = ioloop.IOLoop.instance()
 
     def start(self):
         self.instance.start()
@@ -99,14 +31,6 @@ def main():
     init_db()
     app = ServerApplication()
     app.start()
-
-
-def vmaas_call(endpoint, data):
-    headers = {'Content-type': 'application/json',
-               'Accept': 'application/json'}
-    response = requests.post(VMAAS_URL + endpoint,
-                             data=json.dumps(data), headers=headers)
-    return response.json()
 
 
 if __name__ == '__main__':
